Updated API
The opentelemetry crates have been updated to the latest version which have changes to their API. These changes are also reflected in the API of otlp-logger in that you now have to hold an instance of OtlpLogger to call shutdown. OtlpLogger also does a shutdown via Drop. The metrics layer is also added back again so these are also sent to the otlp endpoint when generated through the specifically named events from tracing-opentelemetry.